requirements.txt
asgiref==3.4.1
certifi==2021.5.30
chardet==4.0.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.5
django-cors-headers==3.7.0
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
docopt==0.6.2
gitignore==0.0.8
gunicorn==20.1.0
idna==3.2
Pillow==8.3.1
psycopg2==2.9.1
psycopg2-binary==2.9.1
PyJWT==2.1.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
static3==0.7.0
urllib3==1.26.6
whitenoise==5.2.0
